+++ android/app/src/main/java/org/kovina/opensend/MainActivity.java
@@ -1,0 +1,5 @@
+package org.kovina.opensend;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}