+++ next.config.ts
@@ -1,7 +1,13 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  output: process.env.CAPACITOR_BUILD ? "export" : undefined,
+  images: {
+    unoptimized: process.env.CAPACITOR_BUILD ? true : undefined,
+  },
+  eslint: {
+    ignoreDuringBuilds: true,
+  },
 };
 
 export default nextConfig;